requirements.txt
asgiref==3.6.0
Django==4.1.7
django-extensions==3.2.1
django-fsm==2.8.1
django-fsm-admin@ git+https://github.com/gadventures/django-fsm-admin.git@fe59a4b09657e011c20e379d262e49c45ba430e7
django-leaflet==0.28.3
django-unfold==0.5.0
factory-boy==3.2.1
Faker==18.3.2
faker-vehicle==0.2.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
django-debug-toolbar==3.8.1
